unify lease statistic reports into one
