+++ leasing/report/lease/lease_statistic_report.py
@@ -1,6 +1,6 @@
 import datetime
 from collections import defaultdict
-from decimal import Decimal
+from decimal import ROUND_HALF_UP, Decimal
 from functools import lru_cache
 
 from django import forms
@@ -9,7 +9,7 @@ from django.utils import formats
 from django.utils.translation import gettext_lazy as _
 from enumfields.drf import EnumField
 
-from leasing.enums import LeaseAreaAttachmentType, LeaseState
+from leasing.enums import LeaseAreaAttachmentType, LeaseState, SubventionType
 from leasing.models import Lease, ServiceUnit
 from leasing.report.lease.common_getters import (
     get_address,
@@ -37,6 +37,25 @@ RESIDENTIAL_INTENDED_USE_IDS = [
     12,
     13,
 ]  # 1 = Asunto, 12 = Asunto, lisärakent., 13 = Asunto 2
+
+
+def get_latest_decision_maker(lease):
+    decision_makers = []
+
+    for decision in lease.decisions.all():
+        decision_makers.append(
+            (
+                decision.decision_maker.name if decision.decision_maker else "",
+                decision.decision_date,
+            )
+        )
+    decision_makers.sort(key=lambda x: x[1] if x[1] else datetime.date.today())
+
+    if len(decision_makers) > 0:
+        latest_decision_maker = decision_makers[-1]
+        return "{} ({})".format(latest_decision_maker[0], latest_decision_maker[1])
+    else:
+        return ""
 
 
 def get_matti_report(obj):
@@ -207,6 +226,175 @@ def get_average_amount_per_area_business(obj):
     )
 
 
+def get_amount_per_area_index_adjusted(lease):
+    volumes = defaultdict(Decimal)
+
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False, amount_per_area__isnull=False
+    ):
+        if basis_of_rent.index:
+            index_ratio = Decimal(basis_of_rent.index.number / 100)
+        else:
+            index_ratio = Decimal(1)
+
+        volumes[basis_of_rent.area_unit] += basis_of_rent.amount_per_area * index_ratio
+
+    return ", ".join(
+        [
+            "{} € / {}".format(
+                formats.number_format(
+                    amounts_per_area.quantize(Decimal(".01"), rounding=ROUND_HALF_UP),
+                    decimal_pos=2,
+                    use_l10n=True,
+                ),
+                area_unit,
+            )
+            for area_unit, amounts_per_area in volumes.items()
+        ]
+    )
+
+
+def get_initial_year_rent(lease):
+    amount = Decimal(0)
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False
+    ):
+        amount += basis_of_rent.calculate_initial_year_rent()
+
+    return amount.quantize(Decimal(".01"), rounding=ROUND_HALF_UP)
+
+
+def get_subsidised_rent(lease):
+    subsidy_amount = Decimal(0)
+    initial_year_rent = Decimal(0)
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False
+    ):
+        initial_year_rent += basis_of_rent.calculate_initial_year_rent()
+        subsidy_amount += basis_of_rent.calculate_subsidy()
+
+    return (initial_year_rent - subsidy_amount).quantize(
+        Decimal(".01"), rounding=ROUND_HALF_UP
+    )
+
+
+def get_subsidy_amount(lease):
+    subsidy_amount = Decimal(0)
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False
+    ):
+        subsidy_amount += basis_of_rent.calculate_subsidy()
+
+    return subsidy_amount.quantize(Decimal(".01"), rounding=ROUND_HALF_UP)
+
+
+def get_subsidy_percent(lease):
+    subsidy_amount = Decimal(0)
+    initial_year_rent = Decimal(0)
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False
+    ):
+        initial_year_rent += basis_of_rent.calculate_initial_year_rent()
+        subsidy_amount += basis_of_rent.calculate_subsidy()
+
+    if (
+        subsidy_amount.compare(Decimal(0)) == 0
+        or initial_year_rent.compare(Decimal(0)) == 0
+    ):
+        return Decimal(0)
+
+    return (subsidy_amount * 100 / initial_year_rent).quantize(
+        Decimal(".01"), rounding=ROUND_HALF_UP
+    )
+
+
+def get_subsidy_amount_per_area(lease):
+    rent_per_area_index_adjusted = Decimal(0)
+    subsidy_amount = Decimal(0)
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False, amount_per_area__isnull=False
+    ):
+        if basis_of_rent.index:
+            index_ratio = Decimal(basis_of_rent.index.number / 100)
+        else:
+            index_ratio = Decimal(1)
+
+        rent_per_area_index_adjusted += basis_of_rent.amount_per_area * index_ratio
+
+        if basis_of_rent.subvention_type == SubventionType.FORM_OF_MANAGEMENT:
+            management_subventions = basis_of_rent.management_subventions.all()
+            if management_subventions:
+                for management_subvention in management_subventions:
+                    subsidy_amount += management_subvention.subvention_amount
+        elif basis_of_rent.subvention_type == SubventionType.RE_LEASE:
+            base_percent = (
+                basis_of_rent.subvention_base_percent
+                if basis_of_rent.subvention_base_percent
+                else 0
+            )
+            graduated_percent = (
+                basis_of_rent.subvention_graduated_percent
+                if basis_of_rent.subvention_graduated_percent
+                else 0
+            )
+            subvention_percent = (
+                1
+                - Decimal(1 - base_percent / 100) * Decimal(1 - graduated_percent / 100)
+            ) * 100
+            subsidy_amount = rent_per_area_index_adjusted * (
+                1 - (subvention_percent / 100)
+            )
+
+    if not subsidy_amount:
+        return Decimal(0)
+
+    return (subsidy_amount - rent_per_area_index_adjusted).quantize(
+        Decimal(".01"), rounding=ROUND_HALF_UP
+    )
+
+
+def get_temporary_discount_percentage(lease):
+    base = Decimal(1)
+    for basis_of_rent in lease.basis_of_rents.filter(
+        archived_at__isnull=True, locked_at__isnull=False
+    ):
+        temporary_subventions = basis_of_rent.temporary_subventions.all()
+        if not temporary_subventions:
+            continue
+
+        for temporary_subvention in temporary_subventions:
+            base *= (100 - temporary_subvention.subvention_percent) / 100
+
+    return (1 - base) * 100
+
+
+def get_temporary_discount_amount(lease):
+    temporary_discount_percentage = get_temporary_discount_percentage(lease)
+    subsidised_rent = get_subsidised_rent(lease)
+
+    return (subsidised_rent / 100 * temporary_discount_percentage).quantize(
+        Decimal(".01"), rounding=ROUND_HALF_UP
+    )
+
+
+def get_subvention_form_of_management(lease):
+    forms_of_management = []
+    for rent in lease.rents.all():
+        for rent_adjustment in rent.rent_adjustments.all():
+            for management_subvention in rent_adjustment.management_subventions.all():
+                if not management_subvention.management:
+                    continue
+                forms_of_management.append(management_subvention.management.name)
+
+    return ", ".join(forms_of_management)
+
+
+def get_form_of_financing(lease):
+    if not lease.financing:
+        return
+    return lease.financing.name
+
+
 class LeaseStatisticReport(AsyncReportBase):
     name = _("Lease statistics report")
     description = _(
@@ -266,6 +454,12 @@ class LeaseStatisticReport(AsyncReportBase):
                 x, include_future_tenants=False, report="Lease statistics report"
             ),
             "width": 40,
+        },
+        # Päättäjä
+        "decision_maker": {
+            "label": _("Decision maker"),
+            "source": get_latest_decision_maker,
+            "width": 20,
         },
         # Start date
         "start_date": {"label": _("Start date"), "format": "date"},
@@ -328,11 +522,78 @@ class LeaseStatisticReport(AsyncReportBase):
             "source": get_average_amount_per_area_business,
             "width": 20,
         },
+        # Yksikköhinta (ind)
+        "amount_per_area_index_adjusted": {
+            "label": _("Unit price (index)"),
+            "source": get_amount_per_area_index_adjusted,
+            "width": 20,
+        },
+        # Alkuvuosivuokra (ind)
+        "initial_year_rent": {
+            "label": _("Initial year rent"),
+            "source": get_initial_year_rent,
+            "format": "money",
+            "width": 13,
+        },
+        # Subventoitu Alkuvuosivuokra (ind)
+        "subsidised_rent": {
+            "label": _("Subsidised rent"),
+            "source": get_subsidised_rent,
+            "format": "money",
+            "width": 13,
+        },
+        # Subventio euroina/vuosi (laskurista)
+        "subsidy_amount": {
+            "label": _("Subsidy amount"),
+            "source": get_subsidy_amount,
+            "format": "money",
+            "width": 13,
+        },
+        # Subventio prosentteina (laskurista)
+        "subsidy_percent": {
+            "label": _("Subsidy percent"),
+            "source": get_subsidy_percent,
+            "format": "percentage",
+            "width": 13,
+        },
+        # Subventoitu eur/k-m2 = Subventoitu yksikköhinta - Yksikköhinta (ind)
+        "subsidy_amount_per_area": {
+            "label": _("Subsidy amount per area"),
+            "source": get_subsidy_amount_per_area,
+            "format": "money",
+            "width": 13,
+        },
+        # Tilapäisalennus subventoidusta alkuvuosivuokrasta % (laskurista)
+        "temporary_discount_percentage": {
+            "label": _("Temporary discount percent"),
+            "source": get_temporary_discount_percentage,
+            "format": "percentage",
+            "width": 13,
+        },
+        # Tilapäisalennus euroa/vuosi (laskurista)
+        "temporary_discount_amount": {
+            "label": _("Temporary discount amount"),
+            "source": get_temporary_discount_amount,
+            "format": "money",
+            "width": 13,
+        },
+        # Hallintamuodon tyyppi (Subventio)
+        "subvention_form_of_management": {
+            "label": _("Form of management (Subvention)"),
+            "source": get_subvention_form_of_management,
+            "width": 20,
+        },
         # Hallintamuoto
         "form_of_management": {
             "label": _("Form of management"),
             "source": get_form_of_management,
             "width": 13,
+        },
+        # Rahoitusmuoto (yhteenvedon tilastotiedoista)
+        "financing": {
+            "label": _("Form of financing"),
+            "source": get_form_of_financing,
+            "width": 20,
         },
         # Erityisasunnot
         "supportive_housing": {
